% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_apply_confidential.R
\name{apply_random_rounding}
\alias{apply_random_rounding}
\title{Apply random rounding to data frame}
\usage{
apply_random_rounding(df, RR_columns, BASE = 3, stable_across_cols = NULL)
}
\arguments{
\item{df}{the data.frame to apply random rounding to.}

\item{RR_columns}{a column or columns of \code{df} to be rounded.}

\item{BASE}{the base to round to. Defaults to 3.}

\item{stable_across_cols}{an optional array of column names. If provided,
these column names are used to generate seeds for the rounding. This
enables consistent or reproducible rounding.}
}
\value{
the data.frame with the columns in \code{RR_columns} renamed with a
\verb{raw_*} prefix, and a matching column with a \verb{conf_*} prefix that contains
randomly rounded values.
}
\description{
Applies random rounding to specified columns.
Creates raw_ and conf_ columns so original values are preserved.
}
