% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_apply_confidential.R
\name{apply_small_count_suppression}
\alias{apply_small_count_suppression}
\title{Apply suppression to data frame}
\usage{
apply_small_count_suppression(
  df,
  suppress_cols,
  threshold,
  count_cols = suppress_cols
)
}
\arguments{
\item{df}{the data.frame to apply suppression to.}

\item{suppress_cols}{the name of the columns to apply suppression to.}

\item{threshold}{the minimum acceptable count. Values less than (but not
equal to) the threshold are suppressed.}

\item{count_cols}{the names of the columns to compare with the threshold.
Optional, defaults to \code{suppress_cols} if not provided. The intention is that
\code{count_cols} contains the counts of unit records.}
}
\value{
the data.frame with the \code{suppress_cols} suppressed where the count
is below the threshold.
}
\description{
Applies suppression to a data frame. Where \code{count_cols} are less than
\code{threshold}, replace \code{suppress_cols} with \code{NA}.
}
\details{
Multiple columns and thresholds can be provided. If multiple \code{suppress_cols}
are given then all receive the same suppression. If multiple \code{count_cols}
are given then suppression is applied if any count is beneath the threshold.
Multiple \code{thresholds} can only be provided if multiple \code{count_cols} are
provided, each count col will use the corresponding threshold.
}
