% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_confidential.R
\name{check_confidentialised_results}
\alias{check_confidentialised_results}
\title{Runs and reports on random rounding, suppression, and handling of zeros
using the most common defaults.}
\usage{
check_confidentialised_results(
  df,
  BASE = 3,
  COUNT_THRESHOLD = 6,
  SUM_THRESHOLD = 20
)
}
\description{
Runs and reports on random rounding, suppression, and handling of zeros
using the most common defaults.
}
