% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_confidential.R
\name{check_graduated_rounding_df}
\alias{check_graduated_rounding_df}
\title{Returns TRUE if the provided column in the dataframe only contains values
that are graduated rounded:}
\usage{
check_graduated_rounding_df(df, column, na.rm = TRUE)
}
\description{
values   base
0-18    3
19     2
20-99    5
100-999   10
1000+  100
}
\details{
If na.rm = TRUE (default) then ignores missing (NA) values.
Warns if all values are missing (NA)
Errors on non-numeric input.
}
