% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_confidential.R
\name{check_random_rounding}
\alias{check_random_rounding}
\title{Check randomness of rounding in data.frame}
\usage{
check_random_rounding(df, raw_col, conf_col, base = 3, print_ratios = FALSE)
}
\arguments{
\item{df}{the data.frame to check rounding of.}

\item{raw_col}{the name of the raw/unrounded column of \code{df}.}

\item{conf_col}{the name of the confidentialised/rounded column of \code{df}.}

\item{base}{the base for the array to be rounded to. Defaults to 3.}

\item{print_ratios}{T/F, whether to display the ratio between the different
rounding differences.}
}
\description{
Check randomness of rounding in data.frame
}
