% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_confidential.R
\name{check_rounding_to_base_array}
\alias{check_rounding_to_base_array}
\title{Filter to a limited number of rows}
\usage{
check_rounding_to_base_array(input_array, base = 3, na.rm = TRUE)
}
\description{
Returns TRUE is the provided array only contains values that are rounded to
the specified base.
}
\details{
If na.rm = TRUE (default) then ignores missing (NA) values.
Else returns FALSE.
Warns if all values are missing (NA)
Errors on non-numeric input.
}
