% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_confidential.R
\name{check_rounding_to_base_df}
\alias{check_rounding_to_base_df}
\title{Returns TRUE if the provided column in the dataframe only contains values
that are rounded to the specified base.}
\usage{
check_rounding_to_base_df(df, column, base = 3, na.rm = TRUE)
}
\description{
If na.rm = TRUE (default) then ignores missing (NA) values.
Warns if all values are missing (NA)
Errors on non-numeric input.
}
