% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_apply_confidential.R
\name{confidentialise_results}
\alias{confidentialise_results}
\title{Apply common confidentiality to long-thin output}
\usage{
confidentialise_results(
  df,
  stable_RR = FALSE,
  sum_RR = FALSE,
  BASE = 3,
  COUNT_THRESHOLD = 6,
  SUM_THRESHOLD = 20
)
}
\arguments{
\item{df}{the data.frame to apply random rounding to. Must be in long-thin
format.}

\item{stable_RR}{T/F, whether to apply stable random rounding using the
non-numeric columns. Defaults to FALSE.}

\item{sum_RR}{T/F, whether to randomly round the \code{sum} column (if present
in \code{df}). Defaults to FALSE. Should be used when the sum is a type of count.
For example, "sum the total number of events for people" is equivalent to
counting all events.}

\item{BASE}{the base to randomly round to. Defaults to 3.}

\item{COUNT_THRESHOLD}{the minimum acceptable count. Where \code{count} or
\code{distinct} are below this threshold then they will be suppressed. Defaults
to 6.}

\item{SUM_THRESHOLD}{the minimum accepted count to output a sum. If either
\code{count} or \code{distinct} are below this threshold, then the sum column will
be suppressed. Defaults to 20.}
}
\value{
the data.frame (still in long-thin format) with the numeric columns
renamed with a \verb{raw_*} prefix, and a matching column with a \verb{conf_*} prefix
that contains the confidentialised values.
}
\description{
Applies the most common confidentiality rules to long-thin output. Designed
to confidentialise in a single command the output from \code{summarise_and_label}
and \code{summarise_and_label_over_lists}.
}
\details{
\itemize{
\item Applies random rounding to columns \code{count} and \code{distinct} if they exist.
\item Applies suppression to columns \code{count} and \code{distinct} if they exist.
\item Applies suppression to column \code{sum} if it exists
\item Ensures consistency between random rounding and suppression (e.g. if a
count is above/below a threshold then it will remain above/below the
threshold after rounding).
}
}
