% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_confidential.R
\name{explore_output_report}
\alias{explore_output_report}
\title{Generate overview reports of summarised dataset for research review.
Visual inspection of these reports as part of checks run during delivery
will help confirm values are reasonable.}
\usage{
explore_output_report(df, output_dir = NA, output_label = NA)
}
\description{
This function writes up to three files to the folder - one for each of
distinct, count, and sum. Each file provides an overview of the range
of values a measure takes.
}
\details{
For example
\itemize{
\item How much does the average value vary across the dataset?
\item How large is a group at its largest and how small at its smallest?
}

It is assumed that a researchers will know what reasonable values are for
each variable and hence will be able to review from such a summary that the
outputs are as expected.

For example, would you expect to observe...
\itemize{
\item the average income to be between $12k and $65k in the results?
\item the number of people to be between 100 and 10,000 in the results?
}

Note that values may fluctuate depending on the subgroups used.
For example, if there are results for both all New Zealand and for only
those people who are recent migrants, then the range of reasonable values
should be wider than if results included only the entire population
or only recent migrants.
}
