% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_summarise.R
\name{has_long_thin_format}
\alias{has_long_thin_format}
\title{Check long-thin format used}
\usage{
has_long_thin_format(df)
}
\arguments{
\item{df}{a data.frame to check for long-thin format}
}
\value{
T/F, whether the data.frame complies with our long-thin format.
}
\description{
For consistency between summarise, confidentialise, and check functionality
we impose a standard format for summarised output. Termed 'long-thin format'.
}
\details{
This format has pairs of 'col' and 'val' columns the describe the group
represented by the data row. Followed by columns for outputs.

Long-thin format takes the form:
\verb{col01, val01, ..., col99, val99, summarised_var, distinct, count,  sum}
Where the last three columns are optional, and can have a \code{raw_} or \code{conf_}
prefix.

Hence we conduct the following checks to confirm a data frame is in this
format: (1) every column name is col_, val_, distinct, count, or sum, and
(2) every col* has a matching val*
}
