% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_apply_confidential.R
\name{randomly_round_vector}
\alias{randomly_round_vector}
\title{Apply random rounded}
\usage{
randomly_round_vector(input_vector, base = 3, seeds = NULL)
}
\arguments{
\item{input_vector}{a numeric array to randomly round.}

\item{base}{the base to round to. defaults to 3.}

\item{seeds}{an optional column of seed values, with length equal to
\code{input_vector}. If provided, these seeds are used to determine the random
rounding - useful for consistent or reproducible rounding.}
}
\value{
the input vector with all values randomly rounded so they are
divisible by \code{base}.
}
\description{
Randomly rounds a numeric vector to the given base.
For example, if base is 3 (default) then:
\itemize{
\item 3 & 6 will be left as is because they are already base 3;
\item 4 & 7 will be rounded down with prob 2/3 and up with prob 1/3;
\item 5 & 8 will be rounded down with prob 1/3 and up with prob 2/3.
}
}
