% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_summarise.R
\name{summarise_and_label}
\alias{summarise_and_label}
\title{Single summarise and label}
\usage{
summarise_and_label(
  df,
  group_by_cols,
  summarise_col,
  make_distinct,
  make_count,
  make_sum,
  clean = "none",
  remove.na.from.groups = TRUE,
  query_path = NA
)
}
\arguments{
\item{df}{a data.frame to summarise. Can be in-memory or remote accessed
with dbplyr.}

\item{group_by_cols}{an array of column names to group the table by.}

\item{summarise_col}{the name of a column to summarise.}

\item{make_distinct}{T/F, whether the output should include a count of
distinct values from \code{summarise_col}.}

\item{make_count}{T/F, whether the output should include a count of values
from \code{summarise_col}.}

\item{make_sum}{T/F, whether the output should include the sum of values
from \code{summarise_col}.}

\item{clean}{hat cleaning, if any, to apply to the summarised column. Must
be one of \verb{\{"none", "na.as.zero", "zero.as.na"\}}. Defaults to \code{"none"}. See
details for explanation.}

\item{remove.na.from.groups}{T/F, whether missing values in the data should
be excluded when grouping for summarisation. For example, if grouping by age
should the output include a row for the records with no age.
Defaults to TRUE.}

\item{query_path}{If provided and data source is SQL will attempt to save a
copy of the SQL code sent to/executed on the database to the provided folder.
Save occurs before execution, hence useful for debugging.}
}
\value{
a local data.frame in long-thin format containing the summarised
results in long-thin format.
}
\description{
Produces a summary of the specified groups with distinct, count, or sum.
Does so in a way that is robust to whether df is a local or a remote table.
Returns a local table in long-thin format.
}
\details{
Used on its own, produces output consistent with \code{dplyr::group_by} followed
by \code{dplyr::summarise}. It's advantage is its consistent output format, and
how it is use with \code{summarise_and_label_over_lists} to produce many
summaries with a single command.

Argument \code{clean} determines how missing values are handled. Missing values
are excluded from the results - they are neither counted nor summed.

The default option \code{"none"} leaves missing values as they appear in the
data. The alternatives \code{"na.as.zero"} and \code{"zero.as.na"} convert between
missing values and zero.

One example where this control is useful is producing average income: Should
it be an average over all people or an average over those people with income?
(As means have additional confidentiality rules, this function produces the
sum/total and the count separately so that confidentiality can be applied
to the numerator and denominator separator if required.)
}
