% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_confidential.R
\name{check_absence_of_zero_counts}
\alias{check_absence_of_zero_counts}
\title{If small non-zero counts are suppressed and zero counts do not appear in the
dataset, then there is a confidentiality risk that true zeros could be
recovered because they are handled differently from small non-zero counts.}
\usage{
check_absence_of_zero_counts(df, conf_count_col, print_on_fail = FALSE)
}
\description{
Returns TRUE if the absence of zero counts does not pose a confidentiality
risk - either there are no suppressed counts OR all combinations of
labels/groups appear in the dataset.
}
\details{
As a summarised dataset may be produced by appending multiple summaries
we run the analysis within each combination.

If print_on_fail = TRUE then at least one combination that is absent from
the dataset is printed.

If a risk is identified (function returns FALSE), there are
two common solutions:
\enumerate{
\item remove all rows that contain suppression of small non-zero counts
\item use expand_to_include_zero_counts to add zero count rows into dataset
}
}
