% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_confidential.R
\name{check_random_rounding}
\alias{check_random_rounding}
\title{Returns TRUE if the provided column in the dataframe has been rounded to the
specified base.}
\usage{
check_random_rounding(
  df,
  raw_col = NA,
  conf_col,
  base = 3,
  print_ratios = FALSE
)
}
\description{
If a raw column is provided then also checks the randomness of the rounding.
\itemize{
\item Warns if a raw/unrounded column is not provided.
\item Warns if rounding does not appear to be random.
\item Warns if difference between raw and rounded columns is too large.
}
}
