% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_confidential.R
\name{check_rounding_to_base_array}
\alias{check_rounding_to_base_array}
\title{Check rounding of array}
\usage{
check_rounding_to_base_array(input_array, base = 3, na.rm = TRUE)
}
\arguments{
\item{input_array}{a numeric array to test for rounding.}

\item{base}{the base for the array to be rounded to. Defaults to 3.}

\item{na.rm}{T/F, whether missing values should be treated as pass or fail.
Defaults to TRUE, so missing values are considered rounded.}
}
\value{
T/F. whether all the values in the array are rounded to \code{base}.
}
\description{
Check rounding of array
}
