% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_confidential.R
\name{check_rounding_to_base_df}
\alias{check_rounding_to_base_df}
\title{Check rounding of column in data.frame}
\usage{
check_rounding_to_base_df(df, column, base = 3, na.rm = TRUE)
}
\arguments{
\item{df}{the data.frame to check rounding of.}

\item{column}{the name of the column of \code{df} to be checked.}

\item{base}{the base for the array to be rounded to. Defaults to 3.}

\item{na.rm}{T/F, whether missing values should be treated as pass or fail.
Defaults to TRUE, so missing values are considered rounded.}
}
\value{
T/F. whether all the values in the array are rounded to \code{base}.
}
\description{
Check rounding of column in data.frame
}
