% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_confidential.R
\name{check_small_count_suppression}
\alias{check_small_count_suppression}
\title{Returns TRUE if the provided column in the dataframe has no values where the
count is less than the threshold.}
\usage{
check_small_count_suppression(
  df,
  suppressed_col,
  threshold,
  count_col = suppressed_col
)
}
\description{
We expect: suppressed_col = NA if count_col < threshold.
Threshold is the smallest acceptable count.
}
